feat: Refactor profile tab to setting tab and update UI
