+++ src/app/(tabs)/setting.tsx
@@ -1,11 +1,11 @@
 import { SafeAreaView } from 'react-native-safe-area-context'
 
-import Mypage from '@/components/mypage/Mypage/Mypage'
+import SettingItem from '@/components/setting/SettingItem/SettingItem'
 
 export default function ProfileScreen() {
   return (
     <SafeAreaView className="mt-[-50px] flex-1 bg-gray-50">
-      <Mypage />
+      <SettingItem />
     </SafeAreaView>
   )
 }
