--- src/components/mypage/Mypage/index.ts
@@ -1,1 +1,0 @@
-export * from './Mypage'
