--- src/components/mypage/MypageHeader/index.ts
@@ -1,1 +1,0 @@
-export * from './MypageHeader'
