+++ src/components/notice/NoticeItem/NoticeItem.tsx
@@ -1,4 +1,3 @@
-import AntDesign from '@expo/vector-icons/AntDesign'
 import { View, Text } from 'react-native'
 
 import { Notice } from '@/types/notice.type'
@@ -9,7 +8,7 @@ interface NoticeItemProps {
 
 export const NoticeItem = ({ item }: NoticeItemProps) => {
   return (
-    <View className="mx-4 mb-3 min-h-[130px] rounded-lg bg-white p-4 shadow-sm">
+    <View className="mx-4 mb-3 min-h-[100px] rounded-lg bg-white p-4 shadow-sm">
       <View className="flex-1 justify-between">
         <Text
           className="text-base font-semibold text-gray-800"
@@ -20,33 +19,19 @@ export const NoticeItem = ({ item }: NoticeItemProps) => {
         <View>
           <View className="mt-2">
             <View className="flex-row items-center justify-between">
-              <View>
-                <Text className="rounded-full bg-slate-100 px-2 py-1 text-xs font-bold text-gray-700">
-                  {item.department}
-                </Text>
-              </View>
-              <View>
-                {item.tags.map((tag) => (
-                  <Text key={tag} className="text-xs text-gray-700">
-                    #{tag}
+              <View className="flex-row items-center justify-between">
+                <View>
+                  <Text className="rounded-full bg-slate-100 px-2 py-1 text-xs font-bold text-gray-700">
+                    {item.department}
                   </Text>
-                ))}
-              </View>
-            </View>
-          </View>
-          <View className="mt-2">
-            <View className="flex-row items-center justify-between">
-              <View className="flex-row items-center gap-1">
-                <AntDesign name="like" size={12} color="black" />
-                <Text className="text-xs text-gray-500">{item.like_count}</Text>
-                <AntDesign name="comment" size={12} color="black" />
-                <Text className="text-xs text-gray-500">
-                  {item.comment_count}
-                </Text>
-                <AntDesign name="book" size={12} color="black" />
-                <Text className="text-xs text-gray-500">
-                  {item.bookmark_count}
-                </Text>
+                </View>
+                <View className="flex-row">
+                  {item.tags.map((tag) => (
+                    <Text key={tag} className="ml-2 text-xs text-gray-700">
+                      #{tag}
+                    </Text>
+                  ))}
+                </View>
               </View>
               <View>
                 <Text className="text-xs text-gray-500">{item.posted_at}</Text>
