+++ src/components/setting/SettingHeader/index.ts
@@ -1,0 +1,1 @@
+export * from './SettingHeader'
