+++ src/components/setting/SettingItem/index.ts
@@ -1,0 +1,1 @@
+export * from './SettingItem'
